pyproject.toml
[project]
name = "ce"
version = "3.6.5"
requires-python = ">=3.11,<4"
description = "PCIC's Climate Explorer (CE)"
authors = [
    {name = "Pacific Climate Impacts Consortium"},
    {name = "Rod Glover", email = "<rglover@uvic.ca>"},
    {name = "James Hiebert", email = "<hiebert@uvic.ca>"},
    {name = "Nik Rados"},
    {name = "Lee Zeman", email = "<lzeman@uvic.ca>"},
    {name = "Jonathan Helfrich"},
    {name = "Ada Sungar"},
    {name = "Cairo Sanders"},
    {name = "Quintin Sparks", email = "<quintins@uvic.ca>"},
    {name = "Basil Veerman"},
    {name = "Eric Yvorchuk"},
]
license = "GPL3"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]
keywords = [
    "sql",
    "api",
    "database",
    "climate",
]
dependencies = [
    "Flask>=3.1.1,<4",

"Flask-SQLAlchemy>=3.0.5,<4",
    "Flask-Cors",
    "shapely>2.0,<3",
    "numpy<2",
    "GDAL>=3.3.0,<3.4",
    "rasterio~=1.3",
    "SQLAlchemy>=2.0.40,<3.0.0",
    "contexttimer==0.3.3",
    "Pint<1",
    "gunicorn==23.0.0",
    "nchelpers @ git+https://github.com/pacificclimate/nchelpers.git@py-sprint-25q1",
    "modelmeta (==2.0.1)",
    "cf-cell-methods==0.1.0",
    "requests>=2.32,<3",
]


[tool.poetry.group.dev.dependencies]
pytest = ">=3.0.0"
python-dateutil = "^2.8.2"
black = "^23.11.0"
pre-commit = "^3.5.0"

[tool.poetry.group.docs.dependencies]
sphinx = "^8.2.3"
sphinx-pyproject = "^0.3.0"
m2r2 = "^0.3.4"
poethepoet = "^0.34.0"

[[tool.poetry.source]]
name = "pcic"
url = "https://pypi.pacificclimate.org/simple"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "online: marks tests that use an online resource (deselect with '-m \"not online\"')",
    "slow: marks tests that are slow (deselect with '-m \"not slow\"')",
]
norecursedirs = ".git site-packages _build *.egg ce/tests/helpers"

[tool.poe.tasks]
document = "sphinx-build doc/source doc/build/html"
pdf = "make latexpdf -C doc"